rename inout.py to fsmedia.py
