+++ scripts/extract.py
@@ -8,7 +8,7 @@ from tqdm import tqdm
 
 from lib.multithreading import pool_process
 from lib.utils import set_system_verbosity
-from scripts.inout import Alignments, Faces, Images, Utils
+from scripts.fsmedia import Alignments, Faces, Images, Utils
 
 class Extract(object):
     """ The extract process. """
